backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
google-search-results
python-dotenv
pydantic
